+++ packages/tinacms/src/toolkit/fields/plugins/mdx-field-plugin/plate/plugins/ui/dropdown.tsx
@@ -1,5 +1,11 @@
-import React, { Fragment } from 'react'
-import { Menu, Transition } from '@headlessui/react'
+import React from 'react'
+import {
+  Menu,
+  MenuButton,
+  MenuItem,
+  MenuItems,
+  Transition,
+} from '@headlessui/react'
 import { ChevronDownIcon } from '@heroicons/react/solid'
 import { classNames } from './helpers'
 
@@ -17,14 +23,13 @@ export function Dropdown({
   return (
     <Menu as="div" className="relative inline-block text-left z-20">
       <div>
-        <Menu.Button className="inline-flex justify-center w-full rounded-md border border-gray-300 shadow-sm px-2 py-1 bg-white text-xs font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-100 focus:ring-blue-500">
+        <MenuButton className="inline-flex justify-center w-full rounded-md border border-gray-300 shadow-sm px-2 py-1 bg-white text-xs font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-100 focus:ring-blue-500">
           {label}
           <ChevronDownIcon className="-mr-1 ml-2 h-4 w-4" aria-hidden="true" />
-        </Menu.Button>
+        </MenuButton>
       </div>
 
       <Transition
-        as={Fragment}
         enter="transition ease-out duration-100"
         enterFrom="transform opacity-0 scale-95"
         enterTo="transform opacity-100 scale-100"
@@ -32,25 +37,25 @@ export function Dropdown({
         leaveFrom="transform opacity-100 scale-100"
         leaveTo="transform opacity-0 scale-95"
       >
-        <Menu.Items className="origin-top-right absolute right-0 mt-2 w-32 max-h-[200px] overflow-y-auto rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 focus:outline-none">
+        <MenuItems className="origin-top-right absolute right-0 mt-2 w-32 max-h-[200px] overflow-y-auto rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 focus:outline-none">
           <div className="py-1">
             {items.map((item) => (
-              <Menu.Item key={item.key}>
-                {({ active }) => (
+              <MenuItem key={item.key}>
+                {({ focus }) => (
                   <button
                     onClick={item.onClick}
                     className={classNames(
-                      active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                      focus ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
                       'block px-4 py-2 text-xs w-full text-right'
                     )}
                   >
                     {item.render}
                   </button>
                 )}
-              </Menu.Item>
+              </MenuItem>
             ))}
           </div>
-        </Menu.Items>
+        </MenuItems>
       </Transition>
     </Menu>
   )