+++ examples/kitchen-sink/CHANGELOG.md
@@ -1,5 +1,14 @@
 # e2e-next
 
+## 0.1.78
+
+### Patch Changes
+
+- Updated dependencies [cf1530d]
+- Updated dependencies [2762994]
+- Updated dependencies [ba5f7a3]
+  - tinacms@2.2.5
+
 ## 0.1.77
 
 ### Patch Changes